feature: add repositories and init their

+++ src/Lab2/Models/Cpus/Cpu.cs
@@ -6,12 +6,12 @@ namespace Itmo.ObjectOrientedProgramming.Lab2.Cpus;
 public class Cpu
 {
     public Cpu(
-        int coreFrequency,
+        double coreFrequency,
         int qtyCore,
         Socket socket,
         DdrStandard ddrStandard,
         int qtyRamSlots,
-        int ramFrequency,
+        double ramFrequency,
         bool graphicCore,
         int tdp,
         int power)
@@ -27,12 +27,12 @@ public class Cpu
         Power = power;
     }
 
-    public int CoreFrequency { get; }
+    public double CoreFrequency { get; }
     public int QtyCore { get; }
     public Socket Socket { get; }
     public DdrStandard DdrStandard { get; }
     public int QtyRamSlots { get; }
-    public int RamFrequency { get; }
+    public double RamFrequency { get; }
     public bool GraphicCore { get; }
     public int Tdp { get; }
     public int Power { get; }
